% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ADMM}
\alias{ADMM}
\title{ADMM algorithm for the estimation of the RSE with the l1 penalty.}
\usage{
ADMM(S_hat, Max_iter, lambda, rho, theta_warm)
}
\arguments{
\item{S_hat}{Multi-taper periodogram estimator.}

\item{Max_iter}{Maximum number of iterations for the ADMM algorithm.}

\item{lambda}{Regularisation parameter (scalar).}

\item{rho}{Augmented Lagrangian parameter}

\item{theta_warm}{Initial value for theta matrix}
}
\value{
A list with components
\item{theta}{Estimate of Inverse spectral density matrix.}
\item{z}{Z estimate from the ADMM algorithm.}
\item{r}{ADMM stopping criteria.}
}
\description{
This function implements the Alternating Direction Method of Multipliers algorithm (ADMM) described in the supplementary material for the estimation of the RSE with the l1 penalty.
}
