% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calc_result}
\alias{calc_result}
\title{Function to extract results from estimation procedure 
 #'}
\usage{
calc_result(res)
}
\arguments{
\item{res}{rds file calcaulted in Estimation.R file}
}
\description{
Function to extract results from estimation procedure 
 #'
}
