% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISE.R
\name{FindNeighbors}
\alias{FindNeighbors}
\title{Find Neighbors Function}
\usage{
FindNeighbors(coordinates, thresh = 0.05)
}
\arguments{
\item{coordinates}{A nx2 matrix with x and y coordinates for each spot, where \code{n} is the number of spots.}

\item{thresh}{Side length (0 to 1) of the square used to define neighborhood range; larger values include more neighbors.}
}
\value{
A nxn logical matrix indicating neighbor relationships between spots.
}
\description{
\code{FindNeighbors} identifies neighboring spots based on their 2D coordinates,
which is common in spatial transcriptomics (ST) data.
Returns a logical matrix indicating neighbors for each spot.
}
