% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{glasso}
\alias{glasso}
\title{Function to obtain glasso estimate for a given scenario and lambda value}
\usage{
glasso(n.stream, chosen_lambda, S_hat, Max_iter)
}
\arguments{
\item{n.stream}{no. of data streams/neurons}

\item{chosen_lambda}{regularization parameter}

\item{S_hat}{multi-trial periodogram estimate}

\item{Max_iter}{Max no. of iterations for the ADMM algorithm}
}
\description{
Function to obtain glasso estimate for a given scenario and lambda value
}
