% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{partial_co}
\alias{partial_co}
\title{Estimation of the partial coherence matrix.}
\usage{
partial_co(theta)
}
\arguments{
\item{theta}{p x p inverse spectral density matrix.}
}
\value{
p x p partial coherence matrix.
}
\description{
This function is used to obtain the partial coherence matrix from an estimate of or the theoretical inverse spectral density matrix.
}
