% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{periodogram}
\alias{periodogram}
\title{Multi-trial Periodogram Function.}
\usage{
periodogram(data, omega, Big_T)
}
\arguments{
\item{data}{Input data of specified structure.}

\item{omega}{Specified frequency of interest.}

\item{Big_T}{Time horizon of the data.}
}
\value{
Periodogram estimate of dimension p x p
}
\description{
This function is used to estimate the multi-taper periodogram for a specified frequency of interest.
}
