% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_ADMM}
\alias{plot_ADMM}
\title{Function to to plot ADMM Stopping Criteria}
\usage{
plot_ADMM(tol, title)
}
\arguments{
\item{tol}{list of stopping criteria vectors for each scenario}

\item{title}{title of the plot (string)}
}
\description{
Function to to plot ADMM Stopping Criteria
}
