% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_eBIC}
\alias{plot_eBIC}
\title{Function to to plot eBIC curves}
\usage{
plot_eBIC(eBIC, title)
}
\arguments{
\item{eBIC}{list of ebic for each scenario}

\item{title}{title of the plot (string)}
}
\description{
Function to to plot eBIC curves
}
