% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_UGM_new}
\alias{plot_UGM_new}
\title{Network plots for Graphical estimator.}
\usage{
plot_UGM_new(pc_list, title_list, points_list)
}
\arguments{
\item{pc_list}{list of partial coherence estimates to be compared.}

\item{titles}{vector of titles for the considered plots.}

\item{points}{vector of points to annotate the network plots.}
}
\value{
Network plots.
}
\description{
This function is used to produce network plots obtained via the Graphical RSE.
}
