% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pre_processing_function}
\alias{pre_processing_function}
\title{Function to restructure Bolding and Franks Data}
\usage{
pre_processing_function(data)
}
\arguments{
\item{data}{(rds) rds file containing raw data}
}
\description{
Function to restructure Bolding and Franks Data
}
