% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISE.R
\name{RISE}
\alias{RISE}
\title{RISE}
\usage{
RISE(spot_express, coordinates, thresh = 0.05, norm = TRUE, cores = 2)
}
\arguments{
\item{spot_express}{A pxn matrix of spot expression, with rows as genes and columns as spots.}

\item{coordinates}{An nx2 matrix with x and y coordinates of each spot, where rows in \code{coordinates} align with columns in \code{spot_express}.}

\item{thresh}{Side length (0 to 1) of the square used to identify neighboring spots; larger values include more neighbors.}

\item{norm}{Logical indicating if \code{spot_express} needs normalized. Default is TRUE, which normalizes \code{spot_express} by scaling factor method}

\item{cores}{Number of CPU cores to use for parallel processing. Default is 2.}
}
\value{
A pxn matrix of spot expression with zeros imputed based on neighboring predictions.
}
\description{
RISE uses predictions from a neighbor-based regression model to impute zeros in spatial transcriptomic (ST) data.
}
