% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_df}
\alias{sample_df}
\title{RISE analysis sample data}
\format{A data frame with 7 rows and 3 variables:
\describe{
  \item{outcomes}{a learning outcome}
  \item{avg_scores}{average score on aligned assessments}
  \item{avg_views}{average views per student of each aligned learning resource}
}}
\usage{
sample_df
}
\description{
Seven learning outcomes, average scores on aligned assessments, and average number of views of each aligned resource.
}
\keyword{datasets}
