% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{spike_density}
\alias{spike_density}
\title{Function to plot firing rate plots from Bolding and Franks Data}
\usage{
spike_density(data, bin_size, a, b, title, c, d)
}
\arguments{
\item{data}{pre-processed data in same format as simulation study}

\item{bin_size}{time bins for data}

\item{a, b}{xlimits for plots}

\item{title}{title of plot}

\item{c, d}{ylimits for plots}
}
\description{
Function to plot firing rate plots from Bolding and Franks Data
}
