% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Trial_freq_Estimator}
\alias{Trial_freq_Estimator}
\title{Function to estimate trial-frequency smoothed periodogram estimate}
\usage{
Trial_freq_Estimator(om, data, Big_T)
}
\arguments{
\item{om}{vector of frequencies}

\item{data}{data of interest in format of nested list of times}

\item{Big_T}{length of time window}
}
\description{
Function to estimate trial-frequency smoothed periodogram estimate
}
